se optimiza el codigo

+++ BlockChain/transactions.cs
@@ -5,7 +5,7 @@
         public string[] add()
         {
             rsaEncryption rsa = new rsaEncryption();
-            int saldo = 100;
+            float saldo = 100.00000000f;
             string x = saldo.ToString();
             string[] z = { rsa.encrypt(x) };
             return z;
